requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyio==4.2.0
    # via azure-core
arrow==1.3.0
    # via django-q
asgiref==3.7.2
    # via
    #   django
    #   django-browser-reload
azure-core==1.29.6
    # via azure-storage-blob
azure-storage-blob==12.19.0
    # via django-distill
blessed==1.20.0
    # via django-q
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==41.0.7
    # via azure-storage-blob
django==4.1.3
    # via
    #   -r requirements.in
    #   django-browser-reload
    #   django-distill
    #   django-picklefield
    #   django-q
    #   django-tailwind
django-browser-reload==1.12.1
    # via django-tailwind
django-distill[microsoft]==3.1.3
    # via -r requirements.in
django-picklefield==3.1
    # via django-q
django-q==1.3.9
    # via -r requirements.in
django-tailwind[reload]==3.8.0
    # via -r requirements.in
idna==3.6
    # via
    #   anyio
    #   requests
isodate==0.6.1
    # via azure-storage-blob
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via arrow
python-dotenv==1.0.0
    # via -r requirements.in
redis==3.5.3
    # via django-q
requests==2.31.0
    # via
    #   azure-core
    #   django-distill
six==1.16.0
    # via
    #   azure-core
    #   blessed
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sqlparse==0.4.4
    # via django
types-python-dateutil==2.8.19.20240106
    # via arrow
typing-extensions==4.9.0
    # via
    #   azure-core
    #   azure-storage-blob
urllib3==2.1.0
    # via requests
wcwidth==0.2.13
    # via blessed
